some cleaning

+++ dotNet5782_3252_2972/BL/IBL.cs
@@ -113,7 +113,7 @@ namespace BlApi
         /// </summary>
         /// <param name="id">the wanted drone's ID</param>
         /// <param name="time">the amount of time the drone was in charging</param>
-        public double DisChargeDrone(int id, float time);
+        public double DisChargeDrone(int id, double time);
 
         /// <summary>
         /// assosiates a drone to a package
